requirements.txt
aenum>=3.1.15
h5py>=3.10.0
matplotlib>=3.5.1
matplotlib-inline>=0.1.3
more-itertools>=8.10.0
numpy
ordered-set>=4.1.0
pandas>=2.1.2
PyLaTeX>=1.4.2
PyYAML>=5.4.1
pyyaml-include>=2.0
regex>=2023.10.3
scipy>=1.8.0
shutilwhich>=1.1.0
sortedcontainers>=2.4.0
soupsieve>=2.3.1
tqdm>=4.66.1
uncertainties>=3.1.7
xmltodict
pybind11
sigmond-scripts
